Change the type of PreviousLogInfo::numRepeated to unsigned long so that it matches the Printf format specifiers when used.

+++ src/common/log.cpp
@@ -149,7 +149,7 @@ struct PreviousLogInfo
     wxLogRecordInfo info;
 
     // the number of times it was already repeated
-    unsigned numRepeated;
+    unsigned long numRepeated;
 };
 
 PreviousLogInfo gs_prevLog;
